+++ src/main/java/com/ispan/projectX/entity/ProductCategory.java
@@ -4,6 +4,9 @@ package com.ispan.projectX.entity;
 import jakarta.persistence.*;
 import jakarta.persistence.GenerationType;
 
+import java.util.ArrayList;
+import java.util.List;
+
 @Entity
 @Table(name = "product_category")
 public class ProductCategory {
@@ -11,7 +14,7 @@ public class ProductCategory {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "category_id")
-    private Long categoryId;
+    private Integer categoryId;
 
     @Column(name = "category_name", nullable = false)
     private String categoryName;
@@ -22,17 +25,17 @@ public class ProductCategory {
     public ProductCategory() {
     }
 
-    public ProductCategory(Long categoryId, String categoryName, String categoryDescription) {
+    public ProductCategory(Integer categoryId, String categoryName, String categoryDescription) {
         this.categoryId = categoryId;
         this.categoryName = categoryName;
         this.categoryDescription = categoryDescription;
     }
 
-    public Long getCategoryId() {
+    public Integer getCategoryId() {
         return categoryId;
     }
 
-    public void setCategoryId(Long categoryId) {
+    public void setCategoryId(Integer categoryId) {
         this.categoryId = categoryId;
     }
